requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
click==8.1.7
    # via -r requirements.in
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.3
    # via jinja2
mistletoe==1.2.1
    # via -r requirements.in
packaging==23.2
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
sqlalchemy==2.0.23
    # via -r requirements.in
typing-extensions==4.9.0
    # via sqlalchemy
